+++ packages/web-components/CHANGELOG.md
@@ -4,7 +4,8 @@ This CHANGELOG.md tracks the updates to the web components package of the CBP de
 
 The React components are wrappers generated from this package and will share the same changes. Projects using React 19 may use the native web components without React wrappers.
 
-## [unpublished] TBD
+## [Unpublished] TBD
+
 
 * Updated the `cbp-slider`:
   * Added range slider capability
@@ -13,6 +14,7 @@ The React components are wrappers generated from this package and will share the
   * Added `filterKeypress` custom event that fires on every keypress when filtering a combobox.
   * Added modifier keys and nativeEvent to the `populateCombobox` event.
 * Updated `cbp-grid` and `cbp-flex` with vertical spacing (based on `gap` by default) when linearized at their responsive breakpoint.
+* Initial creation of `cbp-carousel`, `cbp-carousel-item`, and `cbp-dot-indicator` components.
 
 ## [0.0.1-develop.27] 10-24-2025